move ImmutableInvertedIndex and MutableInvertedIndex into separate files
